+++ swift/Sources/SignalFfi/signal_ffi.h
@@ -175,6 +175,19 @@ static_assert_64bit(alignof(SignalConstPointerPublicKey) == 8);
 typedef const SignalConstPointerPublicKey* SignalType_ConstPointer_SignalConstPointerPublicKey;
 static_assert_64bit(sizeof(SignalType_ConstPointer_SignalConstPointerPublicKey) == 8);
 static_assert_64bit(alignof(SignalType_ConstPointer_SignalConstPointerPublicKey) == 8);
+typedef struct SignalKyberPublicKey SignalKyberPublicKey;
+typedef const SignalKyberPublicKey* SignalType_ConstPointer_SignalKyberPublicKey;
+static_assert_64bit(sizeof(SignalType_ConstPointer_SignalKyberPublicKey) == 8);
+static_assert_64bit(alignof(SignalType_ConstPointer_SignalKyberPublicKey) == 8);
+typedef struct {
+  const SignalKyberPublicKey* raw;
+} SignalConstPointerKyberPublicKey;
+static_assert_64bit(offsetof(SignalConstPointerKyberPublicKey, raw) == 0);
+static_assert_64bit(sizeof(SignalConstPointerKyberPublicKey) == 8);
+static_assert_64bit(alignof(SignalConstPointerKyberPublicKey) == 8);
+typedef const SignalConstPointerKyberPublicKey* SignalType_ConstPointer_SignalConstPointerKyberPublicKey;
+static_assert_64bit(sizeof(SignalType_ConstPointer_SignalConstPointerKyberPublicKey) == 8);
+static_assert_64bit(alignof(SignalType_ConstPointer_SignalConstPointerKyberPublicKey) == 8);
 typedef struct SignalCiphertextMessage SignalCiphertextMessage;
 typedef const SignalCiphertextMessage* SignalType_ConstPointer_SignalCiphertextMessage;
 static_assert_64bit(sizeof(SignalType_ConstPointer_SignalCiphertextMessage) == 8);
@@ -1461,10 +1474,6 @@ typedef struct SignalFingerprint SignalFingerprint;
 typedef const SignalFingerprint* SignalType_ConstPointer_SignalFingerprint;
 static_assert_64bit(sizeof(SignalType_ConstPointer_SignalFingerprint) == 8);
 static_assert_64bit(alignof(SignalType_ConstPointer_SignalFingerprint) == 8);
-typedef struct SignalKyberPublicKey SignalKyberPublicKey;
-typedef const SignalKyberPublicKey* SignalType_ConstPointer_SignalKyberPublicKey;
-static_assert_64bit(sizeof(SignalType_ConstPointer_SignalKyberPublicKey) == 8);
-static_assert_64bit(alignof(SignalType_ConstPointer_SignalKyberPublicKey) == 8);
 typedef struct SignalKyberSecretKey SignalKyberSecretKey;
 typedef const SignalKyberSecretKey* SignalType_ConstPointer_SignalKyberSecretKey;
 static_assert_64bit(sizeof(SignalType_ConstPointer_SignalKyberSecretKey) == 8);
@@ -2896,6 +2905,14 @@ static_assert_64bit(offsetof(SignalBorrowedSliceOfConstPointerPublicKey, length)
 static_assert_64bit(sizeof(SignalBorrowedSliceOfConstPointerPublicKey) == 16);
 static_assert_64bit(alignof(SignalBorrowedSliceOfConstPointerPublicKey) == 8);
 typedef struct {
+  const SignalConstPointerKyberPublicKey* base;
+  size_t length;
+} SignalBorrowedSliceOfConstPointerKyberPublicKey;
+static_assert_64bit(offsetof(SignalBorrowedSliceOfConstPointerKyberPublicKey, base) == 0);
+static_assert_64bit(offsetof(SignalBorrowedSliceOfConstPointerKyberPublicKey, length) == 8);
+static_assert_64bit(sizeof(SignalBorrowedSliceOfConstPointerKyberPublicKey) == 16);
+static_assert_64bit(alignof(SignalBorrowedSliceOfConstPointerKyberPublicKey) == 8);
+typedef struct {
   const SignalConstPointerCiphertextMessage* base;
   size_t length;
 } SignalBorrowedSliceOfConstPointerCiphertextMessage;
@@ -3177,12 +3194,6 @@ typedef struct {
 static_assert_64bit(offsetof(SignalConstPointerFingerprint, raw) == 0);
 static_assert_64bit(sizeof(SignalConstPointerFingerprint) == 8);
 static_assert_64bit(alignof(SignalConstPointerFingerprint) == 8);
-typedef struct {
-  const SignalKyberPublicKey* raw;
-} SignalConstPointerKyberPublicKey;
-static_assert_64bit(offsetof(SignalConstPointerKyberPublicKey, raw) == 0);
-static_assert_64bit(sizeof(SignalConstPointerKyberPublicKey) == 8);
-static_assert_64bit(alignof(SignalConstPointerKyberPublicKey) == 8);
 typedef struct {
   const SignalKyberSecretKey* raw;
 } SignalConstPointerKyberSecretKey;
@@ -3901,6 +3912,15 @@ SignalFfiError* signal_authenticated_chat_connection_set_one_time_ec_pre_keys(
   uint8_t identity_type,
   SignalBorrowedSliceOfu32 pre_key_ids,
   SignalBorrowedSliceOfConstPointerPublicKey pre_key_data
+);
+SignalFfiError* signal_authenticated_chat_connection_set_one_time_kem_pre_keys(
+  SignalCPromisebool* promise,
+  SignalConstPointerTokioAsyncContext async_runtime,
+  SignalConstPointerAuthenticatedChatConnection chat,
+  uint8_t identity_type,
+  SignalBorrowedSliceOfu32 pre_key_ids,
+  SignalBorrowedSliceOfConstPointerKyberPublicKey pre_key_data,
+  SignalBorrowedSliceOfBuffers pre_key_signatures
 );
 SignalFfiError* signal_authenticated_chat_connection_set_push_token_apns(
   SignalCPromisebool* promise,