+++ RELEASE_NOTES.md
@@ -2,3 +2,4 @@ v0.102.3
 
 - New typed APIs:
   - AuthKeysService.setOneTimeEcPreKeys
+  - AuthKeysService.setOneTimeKemPreKeys